remove old downsampling
