+++ tests/test_dataset.py
@@ -1648,172 +1648,6 @@ def test_dataset_conversion():
             )
 
 
-@pytest.fixture(
-    params=[
-        # (offset, size, num_channels, from_mag, max_mag, scale, interpolation_mode, compress, layer type)
-        ((0, 0, 0), (128, 128, 128), 3, "1", "8", None, "default", False, Layer.COLOR_TYPE),
-        ((10, 20, 30), (128, 128, 128), 3, "1", "8", None, "default", False, Layer.COLOR_TYPE),
-        ((10, 20, 30), (128, 128, 128), 3, "1", "8", None, "default", False, Layer.SEGMENTATION_TYPE),
-        #((0, 0, 0), (128, 128, 128), 3, "1", "8", None, "default", True, Layer.COLOR_TYPE), # TODO: find a working configuration for compress
-        #((10, 20, 30), (128, 128, 128), 3, "1", "8", None, "BILINEAR", False, Layer.COLOR_TYPE), # TODO: fix
-        ((10, 20, 30), (128, 128, 128), 3, "1", "8", None, "MEDIAN", False, Layer.COLOR_TYPE),
-        ((10, 20, 30), (128, 128, 128), 3, "1", "8", None, "MODE", False, Layer.COLOR_TYPE),
-        #((10, 20, 30), (128, 128, 128), 3, "1", "8", None, "NEAREST", False, Layer.COLOR_TYPE), # TODO fix
-        ((0, 0, 0), (512, 512, 512), 3, "1", "8", None, "NEAREST", False, Layer.COLOR_TYPE),
-        ((0, 0, 0), (512, 512, 512), 3, "1", "8", None, "BICUBIC", False, Layer.COLOR_TYPE),
-        ((0, 0, 0), (512, 512, 512), 3, "1", "8", None, "BILINEAR", False, Layer.COLOR_TYPE),
-        #((0, 0, 0), (128, 128, 128), 3, "1", "8", None, "BICUBIC", False, Layer.COLOR_TYPE), # TODO fix
-        #((10, 20, 30), (128, 128, 128), 3, "1", "8", None, "BICUBIC", False, Layer.COLOR_TYPE), # TODO fix
-        ((10, 20, 30), (128, 128, 128), 3, "1", "8", None, "MAX", False, Layer.COLOR_TYPE),
-        ((10, 20, 30), (128, 128, 128), 3, "1", "8", None, "MIN", False, Layer.COLOR_TYPE),
-        ((10, 20, 30), (256, 256, 256), 3, "1", "8", None, "default", False, Layer.COLOR_TYPE),
-        ((10, 20, 30), (128, 128, 128), 1, "1", "8", None, "default", False, Layer.COLOR_TYPE),
-        ((10, 20, 30), (128, 128, 128), 3, "2", "8", None, "default", False, Layer.COLOR_TYPE),
-        ((10, 20, 30), (128, 128, 128), 3, "2", "8", (4, 2, 2), "default", False, Layer.COLOR_TYPE),
-    ]
-)
-def generate_dataset(request: Any) -> Tuple:
-    offset, size, num_channels, from_mag, max_mag, scale, interpolation_mode, compress, layer_type = request.param
-
-    ds_path = "./testoutput/larger_wk_dataset/"
-    delete_dir(ds_path)
-
-    ds = WKDataset.create(ds_path, scale=(1, 1, 1))
-    if layer_type == Layer.SEGMENTATION_TYPE:
-        layer = ds.add_layer("layer1", layer_type, num_channels=num_channels, largest_segment_id=1000000000)
-    else:
-        layer = ds.add_layer("layer1", layer_type, num_channels=num_channels)
-    mag = layer.add_mag(str(from_mag), block_len=8, file_len=8)  # use small numbers to better test downsampling
-
-    np.random.seed(1234)
-    if num_channels == 1:
-        write_data = (np.random.rand(size[0], size[1], size[2]) * 255).astype(np.uint8)
-    else:
-        write_data = (np.random.rand(num_channels, size[0], size[1], size[2]) * 255).astype(np.uint8)
-    mag.write(offset=offset, data=write_data)
-
-    return ds_path, Mag(from_mag), Mag(max_mag), scale, interpolation_mode, compress
-
-
-def test_downsampling(generate_dataset: Tuple):
-    origin_path, from_mag, max_mag, scale, interpolation_mode, compress = generate_dataset
-
-    path1 = "./testoutput/downsampling/dataset1/"
-    path2 = "./testoutput/downsampling/dataset2/"
-    path3 = "./testoutput/downsampling/dataset3/"
-
-    delete_dir(path1)
-    delete_dir(path2)
-    delete_dir(path3)
-    copytree(origin_path, path1)
-    copytree(origin_path, path2)
-
-    origin_ds = WKDataset(origin_path)
-    tiff_ds = TiffDataset.create(path3, 1)
-    origin_ds.copy_dataset(tiff_ds)
-    print(origin_ds.layers)
-
-    assert len(origin_ds.layers['layer1'].mags) == 1
-    assert str(from_mag) in origin_ds.layers['layer1'].mags
-
-    # downsample with dataset api
-    #layer2 = TiffDataset(path3).get_layer('layer1')
-    layer2 =WKDataset(path2).get_layer('layer1')
-    layer2.downsample(
-        from_mag=from_mag,
-        max_mag=max_mag,
-        scale=scale,
-        interpolation_mode=interpolation_mode,
-        compress=compress,
-        buffer_edge_len=32
-    )
-
-    # downsample with the old implementation
-    if scale is None:
-        downsample_mags_isotropic(
-            path1,
-            'layer1',
-            from_mag=from_mag,
-            max_mag=max_mag,
-            interpolation_mode=interpolation_mode,
-            compress=compress,
-            buffer_edge_len=32
-        )
-    else:
-        downsample_mags_anisotropic(
-            path1,
-            'layer1',
-            from_mag=from_mag,
-            max_mag=max_mag,
-            scale=scale,
-            interpolation_mode=interpolation_mode,
-            compress=compress,
-            buffer_edge_len=32
-        )
-    layer1 = WKDataset(path1).get_layer('layer1')
-    # update properties manually
-    for res in next(os.walk(Path(path1)/"layer1"))[1]:
-        if res != from_mag.to_layer_name():
-            layer1.setup_mag(Mag(res))
-
-    assert len(layer1.mags) == len(layer2.mags)
-    assert layer1.mags.keys() == layer2.mags.keys()
-
-    for mag in sorted(layer1.mags):
-        data1 = layer1.mags[mag].read()
-        data2 = layer2.mags[mag].read()  # this array might be padded with extra 0s
-        s = data1.shape
-        assert np.array_equal(data1, data2[:s[0], :s[1], :s[2], :s[3]])
-
-
-def test_downsampling_padding():
-    ds_path = "./testoutput/larger_wk_dataset/"
-    delete_dir(ds_path)
-
-    ds = WKDataset.create(ds_path, scale=(1, 1, 1))
-    layer = ds.add_layer("layer1", "segmentation", num_channels=3, largest_segment_id=1000000000)
-    mag1 = layer.add_mag("1", block_len=8, file_len=8)
-    mag2 = layer.add_mag("2", block_len=8, file_len=8)
-
-    # write some random data to mag 1 and mag 2
-    mag1.write(
-        offset=(10, 20, 30),
-        data=(np.random.rand(3, 128, 148, 168) * 255).astype(np.uint8)
-    )
-
-    mag2.write(
-        offset=(5, 10, 15),
-        data=(np.random.rand(3, 64, 74, 84) * 255).astype(np.uint8)
-    )
-
-    assert np.array_equal(mag1.get_view().global_offset, (10, 20, 30))
-    assert np.array_equal(mag1.get_view().size, (128, 148, 168))
-    assert np.array_equal(mag2.get_view().global_offset, (5, 10, 15))
-    assert np.array_equal(mag2.get_view().size, (64, 74, 84))
-
-    layer.downsample(
-        from_mag=Mag(2),
-        max_mag=Mag(8),
-        scale=(4, 2, 2),
-        interpolation_mode="default",
-        compress=False
-    )
-
-    # The data gets padded in a way that it is always possible to divide the offset and the size by 2
-    assert np.array_equal(layer.get_mag("4-8-8").get_view().global_offset, (2, 2, 3))
-    assert np.array_equal(layer.get_mag("2-4-4").get_view().global_offset, (4, 4, 6))
-    assert np.array_equal(mag2.get_view().global_offset, (4, 8, 12))
-    assert np.array_equal(mag1.get_view().global_offset, (8, 16, 24))
-    # The 'end_offset' of the data in mag 1 is: (10, 20, 30) + (128, 148, 168) = (138, 168, 198)
-    # Then 'end_offset' of mag 4-8-8 is: (35, 21, 25)  ->  size=(33, 19, 23)
-    assert np.array_equal(layer.get_mag("4-8-8").get_view().size, (33, 19, 22))
-    assert np.array_equal(layer.get_mag("2-4-4").get_view().size, (66, 38, 44))
-    assert np.array_equal(mag2.get_view().size, (66, 76, 88))
-    assert np.array_equal(mag1.get_view().size, (132, 152, 176))
-
-
-
-
 def test_for_zipped_chunks():
     delete_dir("./testoutput/zipped_chunking_source/")
     delete_dir("./testoutput/zipped_chunking_target/")
@@ -1839,56 +1673,5 @@ def test_for_zipped_chunks():
             executor=executor
         )
 
-
     assert np.array_equal(source_view.read(size=source_view.size) + 50, target_view.read(size=target_view.size))
 
-
-def test_zoom():
-    # This test visualizes why some testcases for downsampling fail
-    factor = 2
-    order = 0
-
-    old_offset = (10, 20, 30)
-    old_data = np.zeros(shape=(128, 128, 128), dtype=np.uint8)
-    for i in range(old_offset[2], old_data.shape[2]):
-        old_data[old_offset[0]:, old_offset[1]:, i] = np.ones(shape=(old_data.shape[0] - old_offset[0], old_data.shape[1] - old_offset[1]), dtype=np.uint8) * (i - old_offset[2] + 100)
-
-    old_downsampled_data = zoom(
-        old_data,
-        1 / factor,
-        output=old_data.dtype,
-        # 0: nearest
-        # 1: bilinear
-        # 2: bicubic
-        order=order,
-        # this does not mean nearest interpolation,
-        # it corresponds to how the borders are treated.
-        mode="nearest",
-        prefilter=True,
-    )
-
-    print(old_downsampled_data.shape)
-
-    # -------------------
-
-    new_offset = (10, 4, 14)
-    new_data = np.zeros(shape=(128, 112, 112), dtype=np.uint8)
-    for i in range(new_offset[2], new_data.shape[2]):
-        new_data[new_offset[0]:, new_offset[1]:, i] = np.ones(shape=(new_data.shape[0] - new_offset[0], new_data.shape[1] - new_offset[1]), dtype=np.uint8) * (i - new_offset[2] + 100)
-
-    new_downsampled_data = zoom(
-        new_data,
-        1 / factor,
-        output=new_data.dtype,
-        # 0: nearest
-        # 1: bilinear
-        # 2: bicubic
-        order=order,
-        # this does not mean nearest interpolation,
-        # it corresponds to how the borders are treated.
-        mode="nearest",
-        prefilter=True,
-    )
-
-    print(new_downsampled_data.shape)
-    assert np.array_equal(old_downsampled_data[:, 8:, 8:], new_downsampled_data)
